# BE-742 | Optimize CalculateTokenOutByTokenIn for balancer pool