--- a/router/usecase/pools/constants.go
+++ b/router/usecase/pools/constants.go
@@ -1,0 +1,7 @@
+package pools
+
+import (
+	"github.com/osmosis-labs/osmosis/osmomath"
+)
+
+var oneDec = osmomath.OneDec()