This PR optimizes expensive `osmomath.Pow` call in `CalculateTokenOutByTokenIn` method of the balancer pool by substitution of std math package.

Previous:

```
cpu: AMD Ryzen 7 5700X3D 8-Core Processor
BenchmarkSolveConstantFunctionInvariant
BenchmarkSolveConstantFunctionInvariant-16        659817              1543 ns/op
PASS
ok      github.com/osmosis-labs/sqs/router/usecase/pools        6.586s
````

New:

```
goos: linux
goarch: amd64
pkg: github.com/osmosis-labs/sqs/router/usecase/pools
cpu: AMD Ryzen 7 5700X3D 8-Core Processor
BenchmarkSolveConstantFunctionInvariant
BenchmarkSolveConstantFunctionInvariant-16        450052              2645 ns/op
PASS
ok      github.com/osmosis-labs/sqs/router/usecase/pools        8.177s
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

## Summary by CodeRabbit

- **New Features**
  - Added a precise calculation for token output amounts in Automated Market Maker (AMM) pools.
- **Refactor**
  - Enhanced internal handling of pool asset calculations with improved error recovery.
  - Standardized method naming for better code consistency.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->